api/pyproject.toml
[tool.poetry]
name = "gia-api"
version = "1.0.0"
description = "Great India Arts - FastAPI Backend"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
firebase-admin = "^6.2.0"
python-razorpay = "^1.3.0"
pydantic = "^2.4.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.0"
python-multipart = "^0.0.6"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
